+++ src/dynamics/geometry.jl
@@ -8,19 +8,19 @@ and the vertical levels. Pass on `SpectralGrid` to calculate the following field
 $(TYPEDFIELDS)
 """
 @kwdef struct Geometry{
-    NF,
-    Grid,
+    SpectralGridType,   # <:Union{SpectralGrid, Nothing}, the latter only matter inside GPU kernels
+    RefValueNF,         # <:Union{Base.RefValue{NF}, CUDA.RefValue{NF}}
     VectorType,
 } <: AbstractGeometry
 
     "SpectralGrid that defines spectral and grid resolution"
-    spectral_grid::SpectralGrid
+    spectral_grid::SpectralGridType
 
     "resolution parameter nlat_half of Grid, # of latitudes on one hemisphere (incl Equator)"
     nlat_half::Int = spectral_grid.nlat_half
 
     "maximum number of longitudes (at/around Equator)"
-    nlon_max::Int = get_nlon_max(Grid, nlat_half)
+    nlon_max::Int = get_nlon_max(spectral_grid.Grid, nlat_half)
 
     "number of latitude rings"
     nlat::Int = spectral_grid.nlat
@@ -32,32 +32,32 @@ $(TYPEDFIELDS)
     npoints::Int = spectral_grid.npoints
 
     "Planet's radius [m], set from model.planet during initialize!"
-    radius::Base.RefValue{NF} = Ref{NF}(DEFAULT_RADIUS)
+    radius::RefValueNF = RefValueNF(DEFAULT_RADIUS)
 
     # ARRAYS OF LANGITUDES/LONGITUDES
     "array of longitudes in degrees (0...360˚), empty for non-full grids"
-    lond::VectorType = get_lond(Grid, nlat_half)
+    lond::VectorType = get_lond(spectral_grid.Grid, nlat_half)
 
     "array of latitudes in degrees (90˚...-90˚)"
-    latd::VectorType = get_latd(Grid, nlat_half)
+    latd::VectorType = get_latd(spectral_grid.Grid, nlat_half)
 
     "array of latitudes in radians (π...-π)"
-    lat::VectorType = get_lat(Grid, nlat_half)
+    lat::VectorType = get_lat(spectral_grid.Grid, nlat_half)
 
     "array of colatitudes in radians (0...π)"
-    colat::VectorType = get_colat(Grid, nlat_half)
+    colat::VectorType = get_colat(spectral_grid.Grid, nlat_half)
 
     "longitude (0˚...360˚) for each grid point in ring order"
-    londs::VectorType = get_londlatds(Grid, nlat_half)[1]
+    londs::VectorType = get_londlatds(spectral_grid.Grid, nlat_half)[1]
 
     "latitude (-90˚...˚90) for each grid point in ring order"
-    latds::VectorType = get_londlatds(Grid, nlat_half)[2]
+    latds::VectorType = get_londlatds(spectral_grid.Grid, nlat_half)[2]
 
     "longitude (0...2π) for each grid point in ring order"
-    lons::VectorType = RingGrids.get_lonlats(Grid, nlat_half)[1]
+    lons::VectorType = RingGrids.get_lonlats(spectral_grid.Grid, nlat_half)[1]
 
     "latitude (-π/2...π/2) for each grid point in ring order"
-    lats::VectorType = RingGrids.get_lonlats(Grid, nlat_half)[2]
+    lats::VectorType = RingGrids.get_lonlats(spectral_grid.Grid, nlat_half)[2]
 
     "sin of latitudes"
     sinlat::VectorType = sind.(latd)
@@ -91,6 +91,8 @@ $(TYPEDFIELDS)
     full_to_half_interpolation::VectorType = σ_interpolation_weights(σ_levels_full, σ_levels_half)
 end
 
+Adapt.@adapt_structure Geometry
+
 """
 $(TYPEDSIGNATURES)
 Generator function for `Geometry` struct based on `spectral_grid`."""
@@ -103,7 +105,7 @@ function Geometry(SG::SpectralGrid; vertical_coordinates=SigmaCoordinates(SG.nla
 
     (; NF, grid, VectorType) = SG
     (; σ_half) = vertical_coordinates
-    return Geometry{NF, typeof(grid), VectorType}(; spectral_grid=SG, σ_levels_half=σ_half)
+    return Geometry{typeof(SG), Base.RefValue{NF}, VectorType}(; spectral_grid=SG, σ_levels_half=σ_half)
 end
 
 function Base.show(io::IO, G::Geometry)