+++ src/physics/albedo.jl
@@ -18,6 +18,8 @@ function Base.show(io::IO, A::OceanLandAlbedo)
     end
 end
 
+Adapt.@adapt_structure OceanLandAlbedo
+
 export DefaultAlbedo
 function DefaultAlbedo(SG::SpectralGrid;
     ocean = OceanSeaIceAlbedo(SG),
@@ -36,6 +38,14 @@ function parameterization!(ij, diagn::DiagnosticVariables, progn, albedo::OceanL
     parameterization!(ij, diagn.physics.land, progn, albedo.land, model)
 end
 
+function variables(::OceanLandAlbedo)
+    return (
+        DiagnosticVariable(name=:albedo, dims=Grid2D(), desc="Albedo", units="1"),
+        DiagnosticVariable(name=:albedo, dims=Grid2D(), desc="Albedo over the ocean", units="1", namespace=:ocean),
+        DiagnosticVariable(name=:albedo, dims=Grid2D(), desc="Albedo over the land", units="1", namespace=:land),
+    )
+end
+
 # single albedo: call separately for ocean and land with the same albedo
 function parameterization!(ij, diagn::DiagnosticVariables, progn, albedo::AbstractAlbedo, model)
     parameterization!(ij, diagn.physics.ocean, progn, albedo, model)
@@ -44,16 +54,18 @@ end
 
 ## GLOBAL CONSTANT ALBEDO
 export GlobalConstantAlbedo
-@kwdef mutable struct GlobalConstantAlbedo{NF} <: AbstractAlbedo
+@kwdef struct GlobalConstantAlbedo{NF} <: AbstractAlbedo
     albedo::NF = 0.3
 end
 
 GlobalConstantAlbedo(SG::SpectralGrid; kwargs...) = GlobalConstantAlbedo{SG.NF}(; kwargs...)
 initialize!(albedo::GlobalConstantAlbedo, ::PrimitiveEquation) = nothing
 parameterization!(ij, diagn, progn, albedo::GlobalConstantAlbedo, model) = albedo!(ij, diagn.albedo, albedo.albedo)
-function albedo!(ij, diagn_albedo::AbstractField2D, albedo::Real)
+@inline function albedo!(ij, diagn_albedo::AbstractArray, albedo::Real)
     diagn_albedo[ij] = albedo
 end
+
+Adapt.@adapt_structure GlobalConstantAlbedo
 
 ## MANUAL ALBEDO
 export ManualAlbedo
@@ -62,20 +74,22 @@ export ManualAlbedo
 Defined so that parameterizations can change the albedo at every time step (e.g. snow cover) without
 losing the information of the original surface albedo. Fields are
 $(TYPEDFIELDS)"""
-struct ManualAlbedo{NF, GridVariable2D} <: AbstractAlbedo
+struct ManualAlbedo{GridVariable2D} <: AbstractAlbedo
     albedo::GridVariable2D
 end
 
-ManualAlbedo(SG::SpectralGrid) = ManualAlbedo{SG.NF, SG.GridVariable2D}(zeros(SG.GridVariable2D, SG.grid))
+ManualAlbedo(SG::SpectralGrid) = ManualAlbedo{SG.GridVariable2D}(zeros(SG.GridVariable2D, SG.grid))
 initialize!(albedo::ManualAlbedo, model::PrimitiveEquation) = nothing
 parameterization!(ij, diagn, progn, albedo::ManualAlbedo, model) = albedo!(ij, diagn.albedo, albedo.albedo)
-function albedo!(ij, diagn_albedo::AbstractField2D, albedo::AbstractField2D)
+@inline function albedo!(ij, diagn_albedo::AbstractArray, albedo::AbstractArray)
     diagn_albedo[ij] = albedo[ij]
 end
 
+Adapt.@adapt_structure ManualAlbedo
+
 ## ALBEDO CLIMATOLOGY
 export AlbedoClimatology
-@kwdef struct AlbedoClimatology{NF, GridVariable2D} <: AbstractAlbedo
+@kwdef struct AlbedoClimatology{GridVariable2D} <: AbstractAlbedo
     "[OPTION] path to the folder containing the albedo file, pkg path default"
     path::String = "SpeedyWeather.jl/input_data"
 
@@ -93,9 +107,9 @@ export AlbedoClimatology
 end
 
 function AlbedoClimatology(SG::SpectralGrid; kwargs...)
-    (; NF, GridVariable2D, grid) = SG
+    (; GridVariable2D, grid) = SG
     albedo = zeros(GridVariable2D, grid)
-    AlbedoClimatology{NF, GridVariable2D}(; albedo, kwargs...)
+    AlbedoClimatology{GridVariable2D}(; albedo, kwargs...)
 end
 
 # set albedo with grid, scalar, function; just define path `albedo.albedo` to grid here
@@ -111,11 +125,14 @@ function initialize!(albedo::AlbedoClimatology, model::PrimitiveEquation)
     end
     ncfile = NCDataset(path)
 
-    a = albedo.file_Grid(ncfile[albedo.varname].var[:, :], input_as=Matrix)
+    a = on_architecture(model.architecture, albedo.file_Grid(ncfile[albedo.varname].var[:, :], input_as=Matrix))
     interpolate!(albedo.albedo, a)
 end
 
 parameterization!(ij, diagn, progn, albedo::AlbedoClimatology, model) = albedo!(ij, diagn.albedo, albedo.albedo)
+
+# For GPU usage just discard the extra information and treat it as a `ManualAlbedo`
+Adapt.adapt_structure(to, albedo::AlbedoClimatology) = adapt(to, ManualAlbedo(albedo.albedo))
 
 ## OceanSeaIceAlbedo
 export OceanSeaIceAlbedo
@@ -129,10 +146,18 @@ OceanSeaIceAlbedo(SG::SpectralGrid; kwargs...) = OceanSeaIceAlbedo{SG.NF}(;kwarg
 initialize!(::OceanSeaIceAlbedo, ::PrimitiveEquation) = nothing
 parameterization!(ij, diagn, progn, albedo::OceanSeaIceAlbedo, model) = albedo!(ij, diagn.albedo, progn.ocean, albedo)
 
-function albedo!(ij, diagn_albedo::AbstractField2D, ocean, albedo::OceanSeaIceAlbedo)
+@inline function albedo!(ij, diagn_albedo::AbstractArray, ocean, albedo::OceanSeaIceAlbedo)
     (; sea_ice_concentration ) = ocean
     (; albedo_ocean, albedo_ice) = albedo
 
     # set ocean albedo linearly between ocean and ice depending on sea ice concentration
     diagn_albedo[ij] = albedo_ocean + sea_ice_concentration[ij] * (albedo_ice - albedo_ocean)
+end
+
+Adapt.@adapt_structure OceanSeaIceAlbedo
+
+function variables(::AbstractAlbedo)
+    return (
+        DiagnosticVariable(name=:albedo, dims=Grid2D(), desc="Albedo", units="1"),
+    )
 end