+++ src/physics/stochastic_physics.jl
@@ -44,3 +44,9 @@ function sppt!(ij, diagn, sppt)
         humid_tend_grid[ij, k] *= R
     end
 end
+
+Adapt.@adapt_structure StochasticallyPerturbedParameterizationTendencies
+
+function variables(::AbstractStochasticPhysics)
+    return ()  # No additional diagnostic variables, modifies existing tendencies
+end